requirements.txt
absl-py==1.4.0
asgiref==3.5.2
asttokens==2.2.1
backcall==0.2.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
colorlog==6.7.0
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
Django==4.1.4
executing==1.2.0
filelock==3.12.0
fonttools==4.39.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
idna==3.4
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipython==8.13.1
jedi==0.18.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
protobuf==4.22.3
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.15.1
pyparsing==3.0.9
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.3.2
zipp==3.15.0